lesson_12/main_app/serializers.py
from rest_framework import serializers
from .models import Post


class PostSerializer(serializers.ModelSerializer):
    class Meta:
        model = Post
        fields = ['id', 'title', 'title_en', 'title_ru', 'title_uz', 'content', 'author']
